[BAEL-4720] Java File.separator vs File.pathSeparator
Fixed code and junits as per review comments
