+++ core-java-modules/core-java-io/src/main/java/com/baeldung/file/separator/FilePathSeparator.java
@@ -6,15 +6,13 @@ import java.util.StringJoiner;
 public class FilePathSeparator {
 
     public static String buildPathUsingString(String[] pathNames) {
-
         return String.join(File.pathSeparator, pathNames);
     }
 
-    public static String buildPathUsingStringJoiner(String path4, String path5) {
+    public static String buildPathUsingStringJoiner(String path1, String path2) {
         StringJoiner joiner = new StringJoiner(File.pathSeparator);
-        joiner.add(path4);
-        joiner.add(path5);
+        joiner.add(path1);
+        joiner.add(path2);
         return joiner.toString();
     }
-
 }
