+++ core-java-modules/core-java-io/src/test/java/com/baeldung/file/separator/FileSeparatorUnitTest.java
@@ -6,19 +6,29 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.EnabledOnOs;
 import org.junit.jupiter.api.condition.OS;
 
-public class FileSeparatorUnitTestForLinux {
+public class FileSeparatorUnitTest {
 
     @Test
-    @EnabledOnOs({ OS.LINUX })
-    public void whenBuildFilePathUsingPathsClass_thenCorrect() {
-        assertEquals("dir1/dir2", FileSeparator.buildFilePathUsingPathsClass("dir1", "dir2"));
-
+    @EnabledOnOs(OS.WINDOWS)
+    public void whenBuildFilePathUsingPathsClass_thenCorrectOnWindows() {
+        assertEquals("dir1\\dir2", FileSeparator.buildFilePathUsingPathsClass("dir1", "dir2"));
     }
 
     @Test
-    @EnabledOnOs({ OS.LINUX })
+    @EnabledOnOs({ OS.LINUX, OS.MAC })
+    public void whenBuildFilePathUsingPathsClass_thenCorrect() {
+        assertEquals("dir1/dir2", FileSeparator.buildFilePathUsingPathsClass("dir1", "dir2"));
+    }
+
+    @Test
+    @EnabledOnOs(OS.WINDOWS)
+    public void whenBuildFilePathUsingFileClass_thenOutputIsAsExpectedOnWindows() {
+        assertEquals("file1\\file2", FileSeparator.buildFilePathUsingFileClass("file1", "file2"));
+    }
+
+    @Test
+    @EnabledOnOs({ OS.LINUX, OS.MAC })
     public void whenBuildFilePathUsingFileClass_thenOutputIsAsExpected() {
         assertEquals("file1/file2", FileSeparator.buildFilePathUsingFileClass("file1", "file2"));
     }
-
 }
